limit GeoTIFF output to float32 data type
git-svn-id: https://www.brockmann-consult.de/svn/os/beam-4.x/trunk@3224 d90d2d48-2b33-0410-a973-e72f85796a02

+++ beam-geotiff/src/test/java/org/esa/beam/dataio/geotiff/GeoTiffBandWriterTest.java
@@ -288,8 +288,8 @@ public class GeoTiffBandWriterTest extends TestCase {
         assertEquals(expSize, _ios.length());
         _ios.seek(offsets[0].getValue());
         for (int j = 0; j < _WIDTH * _HEIGHT; j++) {
-            final double expValue = startValue + j;
-            assertEquals("index at [" + j + "]", expValue, _ios.readDouble(), 1.0e-6);
+            final float expValue = (float) (startValue + j);
+            assertEquals("index at [" + j + "]", expValue, _ios.readFloat(), 1.0e-6);
         }
     }
